fix(PowerUp, GameManager): thiết lập thời gian powerup tồn tại

+++ src/main/java/com/DevChickens/Arkanoid/entities/powerups/ExpandPaddlePowerUp.java
@@ -21,7 +21,7 @@ public class ExpandPaddlePowerUp extends PowerUp {
      * @param x Tọa độ X ban đầu.
      * @param y Tọa độ Y ban đầu.
      */
-    public ExpandPaddlePowerUp(double x, double y, String type, int duration) {
+    public ExpandPaddlePowerUp(double x, double y, String type, long duration) {
         super(x, y, type, duration);
     }
